+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Feature: Add global closing event on `Tooltip`.
 - Fix: Wait until fonts are loaded to trucate titles on `CrowdfundingCard`.
+- Feature: Add `itemsPerPage` prop to use instead of `itemMinWidth` on `Carousel`.
 
 ## [2.87.1] - 2020-08-21